# openapi.yaml
openapi: 3.0.3
info:
  title: SYSmLab API
  version: 1.0.0
  description: |
    API REST mono-laboratório para clientes, pedidos, amostras, resultados,
    laudos verificáveis, inventário, equipamentos e QMS.
    A API apoia controles laboratoriais, mas não concede acreditação.
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
security:
  - bearerAuth: []
...
paths:
...
  /importacao/resultado-analise:
    post:
      tags: [Importação]
      summary: Importa CSV/XLSX como rascunho
      description: Requer Analista ou Gestor; máximo 10 MB e 5.000 linhas.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [arquivo]
              properties:
                arquivo: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /importacao/template:
    get:
      tags: [Importação]
      summary: Baixa template
      parameters:
        - in: query
          name: formato
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        '200':
          description: Arquivo CSV ou XLSX.
          content:
            text/csv:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
...
  /usuarios:
    get:
      tags: [Usuários]
      summary: Lista usuários
      description: Requer Gestor.
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Usuários]
      summary: Cria usuário no Auth e cadastro local
      description: Requer Gestor; a conta administrativa é criada explicitamente aprovada.
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /usuarios/{userId}/perfil:
    put:
      tags: [Usuários]
      summary: Altera perfil
      description: Requer Gestor.
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [perfil]
              properties:
                perfil: { $ref: '#/components/schemas/Role' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
  /usuarios/{userId}/aprovacao:
    put:
      tags: [Usuários]
      summary: Aprova ou bloqueia o acesso de um usuário
      description: >-
        Requer Gestor. Impede auto-bloqueio e bloqueio do último Gestor aprovado;
        a garantia também é aplicada atomicamente no banco.
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [acesso_aprovado]
              properties:
                acesso_aprovado: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Success:
      description: Operação concluída.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessResponse' }
    Created:
      description: Registro criado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessResponse' }
...
    BadRequest:
      description: Payload, parâmetro ou identificador inválido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Regra de estado, concorrência ou domínio impede a operação.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: Limite global ou de confirmação excedido.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  requestBodies:
...
    User:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UserInput' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
        message: { type: string }
        code: { type: string }
        request_id: { type: string, format: uuid }
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:
          nullable: true
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: {} }
            - { type: string }
            - { type: number }
            - { type: boolean }
        count: { type: integer, minimum: 0 }
...
    Role:
      type: string
      enum: [Gestor, Analista, Usuário]
...
    UserInput:
      type: object
      additionalProperties: false
      required: [nome, email, senha, perfil]
      properties:
        nome: { type: string, minLength: 2 }
        email: { type: string, format: email }
        telefone: { type: string, nullable: true }
        senha:
          type: string
          format: password
          minLength: 12
          writeOnly: true
          description: Exige maiúscula, minúscula, número e símbolo.
        perfil: { $ref: '#/components/schemas/Role' }